requirements.txt
absl-py==1.1.0
aioredis==1.3.1
anyio @ file:///C:/ci/anyio_1644463701441/work/dist
arabic-reshaper==2.1.3
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/ci/argon2-cffi-bindings_1644569878360/work
asgi-redis==1.4.3
asgiref==3.2.10
asn1crypto==1.5.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
autobahn==22.3.2
Automat==20.2.0
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
beautifulsoup4 @ file:///C:/ci/beautifulsoup4_1650274792587/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
brotlipy==0.7.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.5.18.1
cffi @ file:///C:/ci_310/cffi_1642682485096/work
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
constantly==15.1.0
cryptography @ file:///C:/ci/cryptography_1652101772384/work
cssselect2==0.6.0
daphne==3.0.2
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Deprecated==1.2.13
Django @ file:///tmp/build/80754af9/django_1601563330930/work
django-cors-headers @ file:///home/conda/feedstock_root/build_artifacts/django-cors-headers_1641858895663/work
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1639677796037/work
entrypoints @ file:///C:/ci/entrypoints_1649926621247/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///tmp/build/80754af9/python-fastjsonschema_1620414857593/work/dist
firebase-admin==5.2.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gcloud==0.18.3
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.1
google-cloud-firestore==2.5.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
h5py==3.7.0
hiredis==2.0.0
html5lib==1.1
httplib2==0.20.4
hyperlink==21.0.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.4
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
incremental==21.3.0
ipykernel @ file:///C:/ci/ipykernel_1647000978151/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1651601994376/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///C:/ci/jedi_1644315425835/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib==1.1.0
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///C:/ci/jsonschema_1650026220037/work
jupyter-client @ file:///C:/ci/jupyter_client_1650534470695/work
jupyter-core @ file:///C:/ci/jupyter_core_1651656285178/work
jupyter-server @ file:///C:/Windows/TEMP/abs_af34539e-be0f-4767-94fb-39ff75b173480cys6cki/croots/recipe/jupyter_server_1654784029941/work
jupyterlab @ file:///opt/conda/conda-bld/jupyterlab_1647445413472/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///C:/ci/jupyterlab_server_1650271783893/work
jwcrypto==1.3.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
lxml==4.8.0
Markdown==3.3.7
MarkupSafe @ file:///C:/ci/markupsafe_1654489871526/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune==0.8.4
msgpack==1.0.3
msgpack-python==0.5.6
nbclassic @ file:///opt/conda/conda-bld/nbclassic_1644943264176/work
nbclient @ file:///C:/ci/nbclient_1650290386732/work
nbconvert @ file:///C:/ci/nbconvert_1649741005454/work
nbformat @ file:///C:/ci/nbformat_1649845101866/work
nest-asyncio @ file:///C:/ci/nest-asyncio_1649829929372/work
notebook @ file:///C:/ci/notebook_1651648658345/work
numpy @ file:///D:/bld/numpy_1640083403086/work
oauth2client==4.1.3
oauthlib==3.2.0
opt-einsum==3.3.0
oscrypto==1.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///D:/bld/pandas_1636223526752/work
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
Paperboy==1.0.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
proto-plus==1.20.5
protobuf==3.19.4
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.14.1
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PyPDF3==1.0.6
Pyrebase4==4.5.0
pyrsistent @ file:///C:/ci/pyrsistent_1636111468851/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-bidi==0.4.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-jwt==3.3.2
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty @ file:///C:/ci_310/pywinpty_1644230983541/work/target/wheels/pywinpty-2.0.2-cp38-none-win_amd64.whl
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1638435185959/work
qrcode==7.3.1
redis==2.10.6
reportlab==3.6.9
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
service-identity==21.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
sniffio @ file:///C:/ci/sniffio_1614030707456/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
svglib==1.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado @ file:///C:/ci/terminado_1644322757089/work
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3 @ file:///C:/ci/urllib3_1650640043075/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///C:/ci/websocket-client_1614804473297/work
Werkzeug==2.1.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt==1.14.1
xgboost==1.5.2
xhtml2pdf==0.2.7
zipp @ file:///C:/ci/zipp_1652274073489/work
zope.interface==5.4.0
